Refactor: 레디스에 key 저장할때 이메일 인증, 토큰인지 prefix 설정
