+++ src/main/java/com/chs/cafeapp/global/redis/constant/RedisKeyPrefix.java
@@ -1,0 +1,6 @@
+package com.chs.cafeapp.global.redis.constant;
+
+public class RedisKeyPrefix {
+  public static final String EMAIL_AUTH = "email auth: ";
+  public static final String TOKEN = "token: ";
+}
